requirements.txt
amqp==2.2.2
appnope==0.1.0
arrow==0.12.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto3==1.5.8
botocore==1.8.22
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
colorlog==3.1.2
decorator==4.1.2
Django==2.0.1
django-s3direct==1.0.4
djangorestframework==3.8.2
django-rest-swagger==2.2.0
docutils==0.14
-e git+git://github.com/hirmeos/entity-fishing-client-python.git@0.4.1#egg=entity_fishing_client
idna==2.6
ipdb==0.11
jedi==0.11.1
jmespath==0.9.3
kombu==4.1.0
lxml==4.2.1
marshmallow==2.15.3
parso==0.1.1
pbr==4.0.2
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
raven==6.4.0
requests==2.18.4
s3transfer==0.1.12
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
zenlog==1.1
